% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response.R
\docType{class}
\name{ApiFileResponse}
\alias{ApiFileResponse}
\title{ApiFileResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApiFileResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{response}}{The raw response from the endpoint.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiFileResponse-new}{\code{ApiFileResponse$new()}}
\item \href{#method-ApiFileResponse-getContent}{\code{ApiFileResponse$getContent()}}
\item \href{#method-ApiFileResponse-getFileName}{\code{ApiFileResponse$getFileName()}}
\item \href{#method-ApiFileResponse-clone}{\code{ApiFileResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiFileResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiFileResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize ApiFileResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiFileResponse$new(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The raw response from the endpoint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiFileResponse-getContent"></a>}}
\if{latex}{\out{\hypertarget{method-ApiFileResponse-getContent}{}}}
\subsection{Method \code{getContent()}}{
Get the http response content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiFileResponse$getContent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiFileResponse-getFileName"></a>}}
\if{latex}{\out{\hypertarget{method-ApiFileResponse-getFileName}{}}}
\subsection{Method \code{getFileName()}}{
Get the filename from the content-disposition http header
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiFileResponse$getFileName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiFileResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiFileResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiFileResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
