% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response.R
\docType{class}
\name{ApiResponse}
\alias{ApiResponse}
\title{ApiResponse}
\format{An \code{R6Class} generator object}
\usage{
ApiResponse
}
\description{
ApiResponse Class
}
\section{Fields}{

\describe{
\item{\code{response}}{The raw response from the endpoint.}
}}

\keyword{datasets}
