% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{blobs_list}
\alias{blobs_list}
\title{[DEPRECATED] List blobs}
\usage{
blobs_list(range = NULL, creation.date = NULL, preload_content = TRUE, ...)
}
\arguments{
\item{range}{(optional) integer}

\item{creation.date}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the blobs in the project
   - `id`: Unique identifier for the blob (UUID)
   - `created_by`: The email of the user who created the blob
   - `creation_date`: Time the blob was created
   - `last_updated`: Time the blob was last updated
   - `filename`: Original filename of the blob
   - `size`: Size of the uploaded blob in bytes
   - `ttl`: Time to live of the blob in seconds
}
\description{
List all blobs in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::blobs_list(
   
   range = NULL, creation.date = NULL
)

# Or provide directly
result <- ubiops::blobs_list(
   
   range = NULL, creation.date = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
