% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{buckets_create}
\alias{buckets_create}
\title{Create bucket}
\usage{
buckets_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, provider, credentials (optional), configuration (optional), ttl (optional), description (optional), labels (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created bucket
  - `id`: Unique identifier for the bucket (UUID)
  - `name`: Name of the bucket
  - `project`: Project name in which the bucket is created
  - `provider`: Provider of the bucket
  - `credentials`: Credentials to connect to the bucket
  - `configuration`: Additional configuration details for the bucket
  - `creation_date`: The date when the bucket was created
  - `description`: Description of the bucket
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `ttl`: Time to live for the files in the bucket
}
\description{
Create a bucket in a project
}
\examples{
\dontrun{
data <- list(
 name = "name",
 provider = "provider",  # one of: [ubiops, google_cloud_storage, amazon_s3, azure_blob_storage] 
 credentials = list(key = "value"),  # (optional)
 configuration = list(key = "value"),  # (optional)
 ttl = 0,  # (optional)
 description = "description",  # (optional)
 labels = list(key = "value")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::buckets_create(
   data
)

# Or provide directly
result <- ubiops::buckets_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
