% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{call_api}
\alias{call_api}
\title{Call API}
\usage{
call_api(url_path, http_method, body = NULL, query_params = NULL,
  content_type = NULL, encode = NULL, UBIOPS_API_TOKEN = NULL,
  UBIOPS_API_URL = NULL, UBIOPS_PROJECT = NULL,
  UBIOPS_TIMEOUT = NULL, UBIOPS_DEFAULT_HEADERS = NULL, ...)
}
\arguments{
\item{url_path}{API endpoint to call, e.g., "status"}

\item{http_method}{HTTP method to use, e.g., "POST"}

\item{body}{body of the request (optional)}

\item{query_params}{query parameters (optional)}

\item{content_type}{content type (optional)}

\item{encode}{encode (optional)}

\item{UBIOPS_API_TOKEN}{(system environment variable) Token to connect to UbiOps API}

\item{UBIOPS_API_URL}{(optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"}

\item{UBIOPS_PROJECT}{(system environment variable) UbiOps project name}

\item{UBIOPS_TIMEOUT}{(optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA}

\item{UBIOPS_DEFAULT_HEADERS}{(optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}

\item{...}{additional parameters to pass to httr GET/POST/PUT/PATCH/HEAD/DELETE function}
}
\value{
Response content from the API
}
\description{
Call an endpoint of the UbiOps API
}
