% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_audit_events_list}
\alias{deployment_audit_events_list}
\title{List audit events for a deployment}
\usage{
deployment_audit_events_list(
  deployment.name,
  action = NULL,
  limit = NULL,
  offset = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{deployment.name}{character}

\item{action}{(optional) character}

\item{limit}{(optional) integer}

\item{offset}{(optional) integer}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the audit events for a deployment
  - `id`: Unique identifier for the audit event (UUID)
  - `date`: The date when the action was performed
  - `action`: Type of action. It can be one of: create, update, delete, info. *info* action denotes that the action does not fall into create, update or delete categories.
  - `user`: Email of the user who performed the action
  - `event`: Description of the event
  - `object_type`: Type of the object on which the action was performed
  - `object_name`: Name of the object on which the action was performed. If the object is deleted at the time of listing audit events, this field is empty.
}
\description{
List all audit events for a deployment including versions
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_audit_events_list(
   deployment.name,
   action = NULL, limit = NULL, offset = NULL
)

# Or provide directly
result <- ubiops::deployment_audit_events_list(
   deployment.name,
   action = NULL, limit = NULL, offset = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
