% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_environment_variables_copy}
\alias{deployment_environment_variables_copy}
\title{Copy deployment environment variable}
\usage{
deployment_environment_variables_copy(
  deployment.name,
  data,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{deployment.name}{character}

\item{data}{named list of: [ source_deployment, source_version (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of the copied variables described by the following fields:
  - `id`: Unique identifier for the environment variable
  - `name`: Variable name
  - `value`: Variable value (will be null for secret variables)
  - `secret`: Boolean that indicates if this variable contains sensitive information
  - `inheritance_type`: Type of parent object that this variable is inherited from. Will be null for copied environment variables.
  - `inheritance_name`: Name of the parent object that this variable is inherited from. Will be null for copied environment variables.
}
\description{
Copy existing environment variables from a source object to the deployment. Variables of the deployment with the same name as ones from the source object will be overwritten with the new value. Only the copied variables are returned.
}
\examples{
\dontrun{
data <- list(
 source_deployment = "source_deployment",
 source_version = "source_version"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_environment_variables_copy(
   deployment.name, data
)

# Or provide directly
result <- ubiops::deployment_environment_variables_copy(
   deployment.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
