% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_environment_variables_delete}
\alias{deployment_environment_variables_delete}
\title{Delete deployment environment variable}
\usage{
deployment_environment_variables_delete(deployment.name, id, ...)
}
\arguments{
\item{deployment.name}{character}

\item{id}{character}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
}
\description{
Delete an environment variable of the deployment
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
ubiops::deployment_environment_variables_delete(
   deployment.name, id
)

# Or provide directly
ubiops::deployment_environment_variables_delete(
   deployment.name, id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
