% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_environment_variables_update}
\alias{deployment_environment_variables_update}
\title{Update deployment environment variable}
\usage{
deployment_environment_variables_update(
  deployment.name,
  id,
  data,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{deployment.name}{character}

\item{id}{character}

\item{data}{named list of: [ name, value, secret ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of variables described by the following fields:
  - `id`: Unique identifier for the environment variable
  - `name`: Variable name
  - `value`: Variable value (will be null for secret variables)
  - `secret`: Boolean that indicates if this variable contains sensitive information
}
\description{
Update an environment variable for the deployment. This cannot be used to update inherited variables; to change an inherited variable for a specific deployment you can create a variable with the same name for the deployment.
}
\examples{
\dontrun{
data <- list(
 name = "name",
 value = "value",
 secret = FALSE
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_environment_variables_update(
   deployment.name, id, data
)

# Or provide directly
result <- ubiops::deployment_environment_variables_update(
   deployment.name, id, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
