% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_requests.R
\name{deployment_requests_batch_get}
\alias{deployment_requests_batch_get}
\title{Retrieve multiple deployment requests}
\usage{
deployment_requests_batch_get(deployment.name, data,
  preload_content = TRUE, ...)
}
\arguments{
\item{deployment.name}{character}

\item{data}{list() - Example: list("request_id_1", "request_id_2")}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of dictionaries containing the details of the retrieved deployment requests with the following fields:
  - `id`: Unique identifier for the deployment request
  - `deployment`: Name of the deployment the request was made to
  - `version`: Name of the version the request was made to
  - `status`: Status of the request. Always 'pending' when initialised, later it can be 'processing', 'failed' or 'completed'.
  - `success`: A boolean value that indicates whether the deployment request was successful. NULL if the request is not yet finished.
  - `time_created`: Server time that the request was made (current time)
  - `time_started`: Server time that the processing of the request was started
  - `time_completed`: Server time that the processing of the request was completed
  - `request_data`: A dictionary containing the data that was sent when the request was created
  - `result`: Deployment request result value. NULL if the request is 'pending', 'processing' or 'failed'.
  - `error_message`: An error message explaining why the request has failed. NULL if the request was successful.
}
\description{
Retrieve multiple deployment requests for the default version of a deployment. If one of the given deployment requests does not exist, an error message is given and no request is returned. A maximum of 250 deployment requests can be retrieved with this method. The deployment requests are NOT returned in the order they are given in.
}
\examples{
\dontrun{
data <- list("request_id_1", "request_id_2")

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_requests_batch_get(
   deployment.name, data
)

# Or provide directly
result <- ubiops::deployment_requests_batch_get(
   deployment.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
