% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_requests.R
\name{deployment_requests_create}
\alias{deployment_requests_create}
\title{Create a direct deployment request}
\usage{
deployment_requests_create(
  deployment.name,
  data,
  timeout = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{deployment.name}{character}

\item{data}{list(key = "value") - Example: list(input_field_1 = "input_value_1", input_field_2 = "input_value_2")}

\item{timeout}{(optional) integer}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created deployment request
  - `id`: Unique identifier for the deployment request
  - `deployment`: Name of the deployment the request was made to
  - `version`: Name of the version the request was made to
  - `success`: A boolean value that indicates whether the deployment request was successful
  - `result`: Deployment request result value. NULL if the request failed.
  - `error_message`: An error message explaining why the request has failed. NULL if the request was successful.
}
\description{
Request a prediction from a deployment. Deployment requests are made for the default version of a deployment. In case of a **blob** type field, the uuid of a previously uploaded blob must be provided.
}
\examples{
\dontrun{
data <- list(input_field_1 = "input_value_1", input_field_2 = "input_value_2")

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_requests_create(
   deployment.name, data,
   timeout = NULL
)

# Or provide directly
result <- ubiops::deployment_requests_create(
   deployment.name, data,
   timeout = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
