% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_requests.R
\name{deployment_requests_get}
\alias{deployment_requests_get}
\title{Get a deployment request}
\usage{
deployment_requests_get(deployment.name, request.id,
  preload_content = TRUE, ...)
}
\arguments{
\item{deployment.name}{character}

\item{request.id}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A dictionary containing the details of the deployment request with the following fields:
  - `id`: Unique identifier for the deployment request
  - `deployment`: Name of the deployment the request was made to
  - `version`: Name of the version the request was made to
  - `status`: Status of the request. Always 'pending' when initialised, later it can be 'processing', 'failed' or 'completed'.
  - `success`: A boolean value that indicates whether the deployment request was successful. NULL if the request is not yet finished.
  - `time_created`: Server time that the request was made (current time)
  - `time_started`: Server time that the processing of the request was started
  - `time_completed`: Server time that the processing of the request was completed
  - `request_data`: A dictionary containing the data that was sent when the request was created
  - `result`: Deployment request result value. NULL if the request is 'pending', 'processing' or 'failed'.
  - `error_message`: An error message explaining why the request has failed. NULL if the request was successful.
  - `created_by`: The email of the user that created the request. In case the request is created by a service, the field will have a "UbiOps" value.
}
\description{
Get a request of the default version of a deployment. With this method, the result of a request may be retrieved.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_requests_get(
   deployment.name, request.id
)

# Or provide directly
result <- ubiops::deployment_requests_get(
   deployment.name, request.id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
