% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_version_environment_variables_create}
\alias{deployment_version_environment_variables_create}
\title{Create deployment version environment variable}
\usage{
deployment_version_environment_variables_create(deployment.name, version,
  data, preload_content = TRUE, ...)
}
\arguments{
\item{deployment.name}{character}

\item{version}{character}

\item{data}{named list of: [ name, value (optional), secret ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of variables described by the following fields:
  - `id`: Unique identifier for the environment variable
  - `name`: Variable name
  - `value`: Variable value (will be null for secret variables)
  - `secret`: Boolean that indicates if this variable contains sensitive information
}
\description{
Create an environment variable for the deployment version. Variables inherited from the project or deployment can be shadowed by creating a variable with the same name.
}
\examples{
\dontrun{
data <- list(
 name = "name",
 value = "value",  # (optional)
 secret = FALSE
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_version_environment_variables_create(
   deployment.name, version, data
)

# Or provide directly
result <- ubiops::deployment_version_environment_variables_create(
   deployment.name, version, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
