% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_versions_create}
\alias{deployment_versions_create}
\title{Create deployment versions}
\usage{
deployment_versions_create(deployment.name, data, preload_content = TRUE,
  ...)
}
\arguments{
\item{deployment.name}{character}

\item{data}{named list of: [ version, language (optional), memory_allocation (optional), maximum_instances (optional), minimum_instances (optional), maximum_idle_time (optional), description (optional), labels (optional), request_retention_time (optional), request_retention_mode (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created version
  - `id`: Unique identifier for the deployment (UUID)
  - `deployment`: Deployment name to which the version is associated
  - `version`: Version name
  - `description`: Description of the version
  - `language`: Language in which the version is provided
  - `status`: The status of the version
  - `active_revision`: Active revision of the version. It is initialised as None since there are no deployment files uploaded for the version yet.
  - `latest_build`: Latest build of the version. It is initialised as None since no build is triggered for the version yet.
  - `memory_allocation`: Reserved memory for the version in MB  
  - `maximum_instances`: Upper bound of number of versions running
  - `minimum_instances`: Lower bound of number of versions running
  - `maximum_idle_time`: Maximum time in seconds a version stays idle before it is stopped
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the version was created
  - `last_updated`: The date when the version was last updated
  - `request_retention_time`: Number of seconds to store requests to the version
  - `request_retention_mode`: Mode of request retention for requests to the version. It can be one of the following: *none*, *metadata* or *full*.
}
\description{
Create a version for a deployment. The first version of a deployment is set as default.
}
\examples{
\dontrun{
data <- list(
 version = "version",
 language = 'python3.7',  # one of: [python3.5, python3.6, python3.7, python3.8, r4.0]  (optional)
 memory_allocation = 0,  # [min: 256; max: 1048576] (optional)
 maximum_instances = 0,  # (optional)
 minimum_instances = 0,  # (optional)
 maximum_idle_time = 0,  # (optional)
 description = "description",  # (optional)
 labels = list(key = "value"),  # (optional)
 request_retention_time = 0,  # [min: 3.6E+3; max: 2.4192E+6] (optional)
 request_retention_mode = 'full'  # one of: [none, metadata, full]  (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_versions_create(
   deployment.name, data
)

# Or provide directly
result <- ubiops::deployment_versions_create(
   deployment.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
