% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_versions_update}
\alias{deployment_versions_update}
\title{Update deployment version}
\usage{
deployment_versions_update(
  deployment.name,
  version,
  data,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{deployment.name}{character}

\item{version}{character}

\item{data}{named list of: [ version (optional), memory_allocation (optional), instance_type (optional), maximum_instances (optional), minimum_instances (optional), maximum_idle_time (optional), description (optional), labels (optional), monitoring (optional), request_retention_time (optional), request_retention_mode (optional), default_notification_group (optional), maximum_queue_size_express (optional), maximum_queue_size_batch (optional), static_ip (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the updated version
  - `id`: Unique identifier for the deployment (UUID)
  - `deployment`: Deployment name to which the version is associated
  - `version`: Version name
  - `description`: Description of the version
  - `language`: Language in which the version is provided
  - `status`: The status of the version
  - `active_revision`: UUID of the active revision of the version. If no deployment files have been uploaded yet, it is None.
  - `latest_build`: UUID of the latest build of the version. If no build has been triggered yet, it is None.
  - `memory_allocation`: (deprecated) Reserved memory for the version in MiB
  - `instance_type`: The reserved instance type for the version
  - `maximum_instances`: Upper bound of number of versions running
  - `minimum_instances`: Lower bound of number of versions running
  - `maximum_idle_time`: Maximum time in seconds a version stays idle before it is stopped
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the version was created
  - `last_updated`: The date when the version was last updated
  - `last_file_upload`: The date when a deployment file was last uploaded for the version
  - `monitoring`: Name of a notification group which contain contacts to send monitoring notifications
  - `default_notification_group`: Name of a notification group which contain contacts to send notifications when requests for the version are completed
  - `request_retention_time`: Number of seconds to store requests to the version
  - `request_retention_mode`: Mode of request retention for requests to the version. It can be one of the following: *none*, *metadata* or *full*.
  - `maximum_queue_size_express`: Maximum number of queued express requests for all instances of this deployment version
  - `maximum_queue_size_batch`: Maximum number of queued batch requests for all instances of this deployment version
  - `has_request_method`: Whether the latest build of the version has a 'request' method
  - `has_requests_method`: Whether the latest build of the version has a 'requests' method
  - `static_ip`: A boolean indicating whether the deployment version should get a static IP
}
\description{
Update a version of a deployment in a project. All necessary fields are validated again. When updating labels, the labels will replace the existing value for labels. Provide the parameter 'monitoring' as the name of a notification group to send monitoring notifications to. A notification will be sent in the case of a failed/recovered request. Pass `null` to switch off monitoring notifications for this version. Provide the parameter 'default_notification_group' as the name of a notification group to send notifications when requests for the version are completed. Pass `null` to switch off request notifications for this version.
}
\examples{
\dontrun{
data <- list(
 version = "version",  # (optional)
 memory_allocation = 0,  # (optional)
 instance_type = "instance_type",  # (optional)
 maximum_instances = 0,  # (optional)
 minimum_instances = 0,  # (optional)
 maximum_idle_time = 0,  # (optional)
 description = "description",  # (optional)
 labels = list(key = "value"),  # (optional)
 monitoring = "monitoring",  # (optional)
 request_retention_time = 0,  # (optional)
 request_retention_mode = "request_retention_mode",  # one of: [none, metadata, full]  (optional)
 default_notification_group = "default_notification_group",  # (optional)
 maximum_queue_size_express = 0,  # (optional)
 maximum_queue_size_batch = 0,  # (optional)
 static_ip = FALSE  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_versions_update(
   deployment.name, version, data
)

# Or provide directly
result <- ubiops::deployment_versions_update(
   deployment.name, version, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
