% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployments_update}
\alias{deployments_update}
\title{Update a deployment}
\usage{
deployments_update(deployment.name, data, preload_content = TRUE, ...)
}
\arguments{
\item{deployment.name}{character}

\item{data}{named list of: [ name (optional), default_version (optional), description (optional), input_fields (optional), output_fields (optional), labels (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the updated deployment
  - `id`: Unique identifier for the deployment (UUID)
  - `name`: Name of the deployment
  - `project`: Project name in which the deployment is defined
  - `input_type`: Type of the input of the deployment
  - `output_type`: Type of the output of the deployment
  - `input_fields`: The list of deployment input fields containing name, data_type and (optional) widget
  - `output_fields`: The list of deployment output fields containing name, data_type and (optional) widget
  - `description`: Description of the deployment
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the deployment was created
  - `last_updated`: The date when the deployment was last updated
  - `default_version`: Default version of the deployment. If it does not have a default version, it is not set.
}
\description{
Update a deployment
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 default_version = "default_version",  # (optional)
 description = "description",  # (optional)
 input_fields = list(  # (optional)
   list(
     name = "name",
     data_type = "data_type",
     widget = widget  # (optional)
   )
 ),
 output_fields = list(  # (optional)
   list(
     name = "name",
     data_type = "data_type",
     widget = widget  # (optional)
   )
 ),
 labels = list(key = "value")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployments_update(
   deployment.name, data
)

# Or provide directly
result <- ubiops::deployments_update(
   deployment.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
