% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{deserialize}
\alias{deserialize}
\title{Deserialize the content of api response}
\usage{
deserialize(resp)
}
\arguments{
\item{resp}{API response}
}
\value{
Deserialized content
}
\description{
Deserialize the content of api response
}
