% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{deserialize_file}
\alias{deserialize_file}
\title{Write file to storage location}
\usage{
deserialize_file(resp, ...)
}
\arguments{
\item{resp}{API response}

\item{...}{additional parameters}
}
\value{
File path
}
\description{
Write file to storage location
}
