% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environment_revisions_get}
\alias{environment_revisions_get}
\title{Get revision}
\usage{
environment_revisions_get(
  environment.name,
  revision.id,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{environment.name}{character}

\item{revision.id}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of a revision
  - `id`: Unique identifier for the revision
  - `environment`: Environment to which the revision is linked
  - `creation_date`: The date when the revision was created
  - `created_by`: The email of the user that uploaded the environment file. In case the revision is created by a service, the field will have a "UbiOps" value.
  - `expired`: A boolean indicating whether the environment file has been deleted for the revision
}
\description{
Retrieve details of a revision of an environment
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::environment_revisions_get(
   environment.name, revision.id
)

# Or provide directly
result <- ubiops::environment_revisions_get(
   environment.name, revision.id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
