% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environments_create}
\alias{environments_create}
\title{Create environments}
\usage{
environments_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, display_name (optional), base_environment, description (optional), labels (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created environment
  - `id`: Unique identifier for the environment
  - `name`: Name of the environment
  - `display_name`: Display name of the environment
  - `base_environment`: Base environment name on which this environment is based
  - `project`: Project name in which the environment is defined
  - `creation_date`: The date when the environment was created
  - `last_updated`: The date when the environment was last updated
  - `description`: Description of the environment
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `gpu_required`: A boolean indicating whether the environment requires GPUs
  - `status`: Status of the environment
  - `implicit`: A boolean indicating whether the environment is implicitly created
  - `hidden`: A boolean indicating whether the environment is hidden
  - `deprecated`: A boolean indicating whether the environment is deprecated
}
\description{
Create a custom environment
}
\examples{
\dontrun{
data <- list(
 name = "name",
 display_name = "display_name",  # (optional)
 base_environment = "base_environment",
 description = "description",  # (optional)
 labels = list(key = "value")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::environments_create(
   data
)

# Or provide directly
result <- ubiops::environments_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
