% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environments_list}
\alias{environments_list}
\title{List environments}
\usage{
environments_list(
  labels = NULL,
  environment.type = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{labels}{(optional) character - Example: "label1:value1,label2:value2"}

\item{environment.type}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the environments
  - `id`: Unique identifier for the environment
  - `name`: Name of the environment
  - `display_name`: Display name of the environment
  - `base_environment`: Base environment name on which this environment is based
  - `project`: Project name in which the environment is defined. It is null for base environments.
  - `creation_date`: The date when the environment was created
  - `last_updated`: The date when the environment was last updated
  - `description`: Description of the environment
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `gpu_required`: A boolean indicating whether the environment requires GPUs
  - `status`: Status of the environment
  - `implicit`: A boolean indicating whether the environment is implicitly created
  - `hidden`: A boolean indicating whether the environment is hidden
  - `deprecated`: A boolean indicating whether the environment is deprecated
}
\description{
Environments can be filtered according to the labels they have by giving labels as a query parameter. Environments that have at least one of the labels on which is filtered, are returned.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::environments_list(
   
   labels = NULL, environment.type = NULL
)

# Or provide directly
result <- ubiops::environments_list(
   
   labels = NULL, environment.type = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
