% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environments_usage}
\alias{environments_usage}
\title{List usage of environment}
\usage{
environments_usage(
  environment.name,
  environment.type = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{environment.name}{character}

\item{environment.type}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the deployment versions
  - `id`: Unique identifier for the deployment version (UUID)
  - `deployment`: Deployment name to which the version is associated
  - `version`: Version name
  - `environment`: Environment of the version
  - `environment_display_name`: Human readable name of the environment
  - `status`: The status of the version
}
\description{
List the deployment versions used by an environment
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::environments_usage(
   environment.name,
   environment.type = NULL
)

# Or provide directly
result <- ubiops::environments_usage(
   environment.name,
   environment.type = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
