% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_and_exports.R
\name{exports_list}
\alias{exports_list}
\title{List exports}
\usage{
exports_list(status = NULL, preload_content = TRUE, ...)
}
\arguments{
\item{status}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the exports in the project
  - `id`: Unique identifier for the export (UUID)
  - `creation_date`: Time the export was created
  - `status`: The status of the export
  - `error_message`: The error message in case of a failure
  - `size`: Size of the export in bytes
}
\description{
List all exports in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::exports_list(
   
   status = NULL
)

# Or provide directly
result <- ubiops::exports_list(
   
   status = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
