% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{expressions_evaluate}
\alias{expressions_evaluate}
\title{Evaluate expression}
\usage{
expressions_evaluate(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ expression, input_fields, request_data (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
}
\description{
Evaluate a pipeline version operator expression.
}
\examples{
\dontrun{
data <- list(
 expression = "expression",
 input_fields = list(  # (optional)
   list(
     name = "name",
     data_type = "data_type"
   )
 ),
 request_data = list(key = "value")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::expressions_evaluate(
   data
)

# Or provide directly
result <- ubiops::expressions_evaluate(
   data,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
