% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files_list}
\alias{files_list}
\title{List files}
\usage{
files_list(
  bucket.name,
  prefix = NULL,
  delimiter = NULL,
  continuation.token = NULL,
  limit = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{bucket.name}{character}

\item{prefix}{(optional) character}

\item{delimiter}{(optional) character}

\item{continuation.token}{(optional) character}

\item{limit}{(optional) integer}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A dictionary containing the details of files and prefixes in the bucket
  - `continuation_token`: Next token to get the next set of files
  - `files`: A list of dictionaries containing the details of the files. It contains the file name ('file'), size of the file ('size') and the creation time of the file ('time_created').
  - `prefixes`: A list of directories
}
\description{
List files in a bucket. The maximum number of files that can be listed at once is 1000.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::files_list(
   bucket.name,
   prefix = NULL, delimiter = NULL, continuation.token = NULL, limit = NULL
)

# Or provide directly
result <- ubiops::files_list(
   bucket.name,
   prefix = NULL, delimiter = NULL, continuation.token = NULL, limit = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
