% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files_start_multipart_upload}
\alias{files_start_multipart_upload}
\title{Start multipart upload}
\usage{
files_start_multipart_upload(
  bucket.name,
  file,
  data = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{bucket.name}{character}

\item{file}{character}

\item{data}{(optional) list(key = "value")}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 - `upload_id`: ID of the upload for the file
  - `provider`: Provider of the bucket where the file will be uploaded
}
\description{
Start a multipart upload for a file
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::files_start_multipart_upload(
   bucket.name, file,
   data = NULL
)

# Or provide directly
result <- ubiops::files_start_multipart_upload(
   bucket.name, file,
   data = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
