% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files_upload}
\alias{files_upload}
\title{Upload a file}
\usage{
files_upload(
  bucket.name,
  file,
  upload.id = NULL,
  part.number = NULL,
  data = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{bucket.name}{character}

\item{file}{character}

\item{upload.id}{(optional) character}

\item{part.number}{(optional) character}

\item{data}{(optional) list(key = "value")}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 - `url`: A url which can be used to upload the file to bucket. Make a PUT request to this url with the file content to upload the file.
  - `provider`: Provider of the bucket where the file will be uploaded
}
\description{
Generate a signed url to upload a file. Request body should be an empty dictionary.  Note: When using the url generated by this endpoint for Azure Blob Storage, the following headers must be added to the upload request to Azure Blob Storage: - `x-ms-version`: '2020-04-08' - `x-ms-blob-type`: 'BlockBlob'
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::files_upload(
   bucket.name, file,
   upload.id = NULL, part.number = NULL, data = NULL
)

# Or provide directly
result <- ubiops::files_upload(
   bucket.name, file,
   upload.id = NULL, part.number = NULL, data = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
