% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_and_exports.R
\name{imports_download}
\alias{imports_download}
\title{Download an import}
\usage{
imports_download(import.id, preload_content = FALSE, ...)
}
\arguments{
\item{import.id}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the output file is loaded and written to the storage location defined by UBIOPS_TEMP_FOLDER_PATH. When FALSE, unprocessed API response object is returned. - Default = FALSE}

\item{...}{UBIOPS_TEMP_FOLDER_PATH (system environment variable) Storage location for output file when `preload_content` is set to TRUE - Default = current working directory
UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
}
\description{
Download an import in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::imports_download(
   import.id
)

# Or provide directly
result <- ubiops::imports_download(
   import.id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

# Save file manually
file_name <- result$getFileName()

output <- file(file.path(getwd(), file_name), "wb")
readr::write_file(result$getContent(), output)
close(output)

# Or save file directly in temp folder
# by using environment variables or providing directly
Sys.setenv("UBIOPS_TEMP_FOLDER_PATH" = getwd())
output_path <- ubiops::imports_download(
  import.id,
  preload_content=TRUE
)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
