% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_and_exports.R
\name{imports_get}
\alias{imports_get}
\title{Get an import}
\usage{
imports_get(import.id, preload_content = TRUE, ...)
}
\arguments{
\item{import.id}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 - `id`: Unique identifier for the import (UUID)
  - `status`: Status of the import
  - `error_message`: The error message in case of a failure
  - `creation_date`: The date when the import was created
  - `size`: Size of the import in bytes
  - `deployments`: Dictionary of the deployments in the import
  - `pipelines`: Dictionary of the pipelines in the import
  - `environment_variables`: Dictionary of the environment variables in the import
  - `environments`: Dictionary of the environments in the import
}
\description{
Get the details of an import in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::imports_get(
   import.id
)

# Or provide directly
result <- ubiops::imports_get(
   import.id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
