% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{instance_types_list}
\alias{instance_types_list}
\title{List instance types}
\usage{
instance_types_list(preload_content = TRUE, ...)
}
\arguments{
\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the instance type
  - `id`: Unique identifier for the instance type (UUID)
  - `name`: Name of the deployment instance type
  - `display_name`: Readable name of the deployment instance type
  - `memory_allocation`: Integer indicating memory allocation for this instance type (Mi)
  - `cpu_allocation`: Float indicating vCPU allocation for this instance type
  - `gpu_allocation`: Integer indicating number of GPU cores for this instance type
  - `storage_allocation`: Integer indicating the maximum storage that can be used by this instance type (MB)
  - `credit_rate`: Credits used per hour for this instance type
}
\description{
Get list of available deployment instance types for a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::instance_types_list(
   
)

# Or provide directly
result <- ubiops::instance_types_list(
   
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
