% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_create}
\alias{metrics_create}
\title{Create metrics}
\usage{
metrics_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, description (optional), metric_type, unit (optional), labels (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created metric
  - `id`: Unique identifier for the metric
  - `name`: Name of the metric
  - `description`: Description of the metric
  - `creation_date`: The date when the metric was created
  - `last_updated`: The date when the metric was last updated
  - `custom`: A boolean indicating whether the metric is custom
  - `metric_type`: Type of the metric
  - `unit`: Unit of the metric
  - `labels`: A list of labels that can be used to get data points containing the metric
}
\description{
Create a custom metric. The name must start with *custom.*.
}
\examples{
\dontrun{
data <- list(
 name = "name",
 description = "description",  # (optional)
 metric_type = "metric_type",
 unit = "unit",  # (optional)
 labels = list("value-1", "value-2")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::metrics_create(
   data
)

# Or provide directly
result <- ubiops::metrics_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
