% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{notification_groups_create}
\alias{notification_groups_create}
\title{Create notification groups}
\usage{
notification_groups_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, contacts ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created notification group
  - `id`: Unique identifier for the notification group (UUID)
  - `name`: Name of the notification group
  - `contacts`: A list of contacts in the notification group
}
\description{
Create a notification group by defining a name and a list of contacts
}
\examples{
\dontrun{
data <- list(
 name = "name",
 contacts = list(  # (optional)
   list(
     type = "type",
     configuration = list(key = "value")
   )
 )
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::notification_groups_create(
   data
)

# Or provide directly
result <- ubiops::notification_groups_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
