% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{notification_groups_get}
\alias{notification_groups_get}
\title{Get notification group}
\usage{
notification_groups_get(notification.group.name, preload_content = TRUE, ...)
}
\arguments{
\item{notification.group.name}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of a notification group
  - `id`: Unique identifier for the notification group (UUID)
  - `name`: Name of the notification group
  - `contacts`: A list of contacts in the notification group
}
\description{
Retrieve details of a single notification group in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::notification_groups_get(
   notification.group.name
)

# Or provide directly
result <- ubiops::notification_groups_get(
   notification.group.name,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
