% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{organization_users_list}
\alias{organization_users_list}
\title{List the users in an organization}
\usage{
organization_users_list(organization.name, preload_content = TRUE, ...)
}
\arguments{
\item{organization.name}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 List of details of users
  - `id`: Unique identifier for the user (UUID)
  - `email`: Email of the user
  - `name`: Name of the user
  - `surname`: Surname of the user
  - `admin`: Boolean value indicating whether the user is an admin of the organization or not
}
\description{
List users and their details in an organization
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::organization_users_list(
   organization.name
)

# Or provide directly
result <- ubiops::organization_users_list(
   organization.name,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
