% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{organizations_create}
\alias{organizations_create}
\title{Create organizations}
\usage{
organizations_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, subscription (optional), subscription_end_date (optional), voucher (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created organization
  - `id`: Unique identifier for the organization (UUID)
  - `name`: Name of the organization
  - `creation_date`: Date and time the organization was created
}
\description{
Create a new organization. When a user creates an organization, s/he will automatically become an organization admin.
}
\examples{
\dontrun{
data <- list(
 name = "name",
 subscription = "subscription",  # (optional)
 subscription_end_date = subscription_end_date,  # (optional)
 voucher = "voucher"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::organizations_create(
   data
)

# Or provide directly
result <- ubiops::organizations_create(
   data,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
