% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{organizations_get}
\alias{organizations_get}
\title{Get details of an organization}
\usage{
organizations_get(organization.name, preload_content = TRUE, ...)
}
\arguments{
\item{organization.name}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the organization
  - `id`: Unique identifier for the organization (UUID) 
  - `name`: Name of the organization 
  - `creation_date`: Time the organization was created 
  - `subscription`: Name of the subscription of the organization
}
\description{
Get the details of an organization
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::organizations_get(
   organization.name
)

# Or provide directly
result <- ubiops::organizations_get(
   organization.name,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
