% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_requests.R
\name{pipeline_requests_delete}
\alias{pipeline_requests_delete}
\title{Delete a pipeline request}
\usage{
pipeline_requests_delete(pipeline.name, request.id, ...)
}
\arguments{
\item{pipeline.name}{character}

\item{request.id}{character}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
}
\description{
Delete a request for the default version of a pipeline. This action deletes all the deployment requests in the pipeline.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
ubiops::pipeline_requests_delete(
   pipeline.name, request.id
)

# Or provide directly
ubiops::pipeline_requests_delete(
   pipeline.name, request.id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
