% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_version_object_attachments_delete}
\alias{pipeline_version_object_attachments_delete}
\title{Delete object attachment}
\usage{
pipeline_version_object_attachments_delete(attachment.id, pipeline.name,
  version, ...)
}
\arguments{
\item{attachment.id}{character}

\item{pipeline.name}{character}

\item{version}{character}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
}
\description{
Delete an attachment in a pipeline version. The referenced and original objects of the attachment still exist in the pipeline version, only the link between them is deleted.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
ubiops::pipeline_version_object_attachments_delete(
   attachment.id, pipeline.name, version
)

# Or provide directly
ubiops::pipeline_version_object_attachments_delete(
   attachment.id, pipeline.name, version,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
