% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_version_object_attachments_list}
\alias{pipeline_version_object_attachments_list}
\title{List object attachments}
\usage{
pipeline_version_object_attachments_list(
  pipeline.name,
  version,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{pipeline.name}{character}

\item{version}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the attachments in the pipeline
  - `destination_name`: Name of the destination pipeline object
  - `sources`: A list of dictionaries containing the source object(s) and mapping of the fields. One attachment can have multiple source objects.
}
\description{
List all attachments in a pipeline version
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_version_object_attachments_list(
   pipeline.name, version
)

# Or provide directly
result <- ubiops::pipeline_version_object_attachments_list(
   pipeline.name, version,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
