% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_version_objects_create}
\alias{pipeline_version_objects_create}
\title{Create pipeline object}
\usage{
pipeline_version_objects_create(
  pipeline.name,
  version,
  data,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{pipeline.name}{character}

\item{version}{character}

\item{data}{named list of: [ name, reference_name, version (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created pipeline object
  - `id`: Unique identifier for the pipeline object (UUID)
  - `name`: Name of the pipeline object
  - `reference_name`: Name of the object it will reference
  - `version`: Version name of reference object
}
\description{
Create a pipeline object for a pipeline version. The pipeline object that is added is a reference to the real object. In this way, multiple references to the same object may be added to a pipeline version. The reference_name refers to the deployment name and the version is the version of the deployment which will be added to the pipeline version as an object.
}
\examples{
\dontrun{
data <- list(
 name = "name",
 reference_name = "reference_name",
 version = "version"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_version_objects_create(
   pipeline.name, version, data
)

# Or provide directly
result <- ubiops::pipeline_version_objects_create(
   pipeline.name, version, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
