% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_version_objects_update}
\alias{pipeline_version_objects_update}
\title{Update pipeline object}
\usage{
pipeline_version_objects_update(name, pipeline.name, version, data,
  preload_content = TRUE, ...)
}
\arguments{
\item{name}{character}

\item{pipeline.name}{character}

\item{version}{character}

\item{data}{named list of: [ name (optional), version (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the updated pipeline object
  - `id`: Unique identifier for the pipeline object (UUID)
  - `name`: Name of the pipeline object
  - `reference_name`: Name of the object it references
  - `version`: Version name of reference object
}
\description{
Update a pipeline object. It is not possible to update the reference_name. All necessary fields are validated again.
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 version = "version"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_version_objects_update(
   name, pipeline.name, version, data
)

# Or provide directly
result <- ubiops::pipeline_version_objects_update(
   name, pipeline.name, version, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
