% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_requests.R
\name{pipeline_version_requests_batch_get}
\alias{pipeline_version_requests_batch_get}
\title{Retrieve multiple pipeline version requests}
\usage{
pipeline_version_requests_batch_get(pipeline.name, version, data,
  preload_content = TRUE, ...)
}
\arguments{
\item{pipeline.name}{character}

\item{version}{character}

\item{data}{list() - Example: list("request_id_1", "request_id_2")}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of dictionaries containing the details of the retrieved pipeline requests with the following fields:
   - `id`: Unique identifier for the pipeline request
   - `pipeline`: Name of the pipeline for which the request is made
   - `version`: Name of the pipeline version for which the request was made
   - `status`: Status of the request. Can be 'pending', 'processing', 'failed' or 'completed'.
   - `success`: A boolean value that indicates whether the pipeline request was successful. NULL if the request is not yet finished.
   - `time_created`: Server time that the request was made (current time)
   - `time_started`: Server time that the processing of the request was started
   - `time_completed`: Server time that the processing of the request was completed
   - `request_data`: A dictionary (structured input type) or string (plain input type) containing the data that was sent when the request was created
   - `result`: A dictionary (structured output type) or string (plain output type) containing the data connected to the pipeline end
   - `deployment_requests`: A list of requests to the deployments in the pipeline. With the deployment request ids provided in this list, it's possible to collect the results of the deployment requests separately.
   - `error_message`: An error message explaining why the request has failed. NULL if the request was successful.
}
\description{
Retrieve multiple requests for a pipeline version. If one of the given pipeline requests does not exist, an error message is given and no request is returned. A maximum of 100 pipeline version requests can be retrieved with this method. The pipeline version requests are NOT returned in the order they are given in.
}
\examples{
\dontrun{
data <- list("request_id_1", "request_id_2")

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_version_requests_batch_get(
   pipeline.name, version, data
)

# Or provide directly
result <- ubiops::pipeline_version_requests_batch_get(
   pipeline.name, version, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
