% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_requests.R
\name{pipeline_version_requests_list}
\alias{pipeline_version_requests_list}
\title{List pipeline version requests}
\usage{
pipeline_version_requests_list(
  pipeline.name,
  version,
  status = NULL,
  success = NULL,
  limit = NULL,
  offset = NULL,
  sort = NULL,
  start.date = NULL,
  end.date = NULL,
  search.id = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{pipeline.name}{character}

\item{version}{character}

\item{status}{(optional) character}

\item{success}{(optional) character ("true"|"false")}

\item{limit}{(optional) integer}

\item{offset}{(optional) integer}

\item{sort}{(optional) character}

\item{start.date}{(optional) character}

\item{end.date}{(optional) character}

\item{search.id}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of dictionaries containing the details of the pipeline version requests with the following fields:
   - `id`: Unique identifier for the pipeline version request
   - `pipeline`: Name of the pipeline for which the request is made
   - `version`: Name of the pipeline version for which the request was made
   - `status`: Status of the request
   - `success`: A boolean value that indicates whether the pipeline version request was successful. NULL if the request is not yet finished.
   - `time_created`: Server time that the request was made (current time)
   - `time_started`: Server time that the processing of the request was started
   - `time_completed`: Server time that the processing of the request was completed
}
\description{
List all batch requests for a pipeline version
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_version_requests_list(
   pipeline.name, version,
   status = NULL, success = NULL, limit = NULL, offset = NULL, sort = NULL, start.date = NULL, end.date = NULL, search.id = NULL
)

# Or provide directly
result <- ubiops::pipeline_version_requests_list(
   pipeline.name, version,
   status = NULL, success = NULL, limit = NULL, offset = NULL, sort = NULL, start.date = NULL, end.date = NULL, search.id = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
