% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_versions_create}
\alias{pipeline_versions_create}
\title{Create pipeline versions}
\usage{
pipeline_versions_create(pipeline.name, data, preload_content = TRUE,
  ...)
}
\arguments{
\item{pipeline.name}{character}

\item{data}{named list of: [ version, description (optional), labels (optional), request_retention_time (optional), request_retention_mode (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created pipeline version
  - `id`: Unique identifier for the pipeline version (UUID)
  - `pipeline`: Name of the pipeline to which the version is associated
  - `version`: Name of the version of the pipeline
  - `description`: Description of the pipeline version
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the pipeline version was created
  - `last_updated`: The date when the pipeline version was last updated
  - `request_retention_time`: Number of seconds to store requests to the pipeline version
  - `request_retention_mode`: Mode of request retention for requests to the pipeline version. It can be one of the following: *none*, *metadata* or *full*.
}
\description{
Create a version for a pipeline. The first version of a pipeline is set as default.
}
\examples{
\dontrun{
data <- list(
 version = "version",
 description = "description",  # (optional)
 labels = list(key = "value"),  # (optional)
 request_retention_time = 0,  # [min: 3.6E+3; max: 2.4192E+6] (optional)
 request_retention_mode = 'full'  # one of: [none, metadata, full]  (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_versions_create(
   pipeline.name, data
)

# Or provide directly
result <- ubiops::pipeline_versions_create(
   pipeline.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
