% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_versions_get}
\alias{pipeline_versions_get}
\title{Get pipeline version}
\usage{
pipeline_versions_get(pipeline.name, version, preload_content = TRUE, ...)
}
\arguments{
\item{pipeline.name}{character}

\item{version}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the pipeline version
  - `id`: Unique identifier for the pipeline version (UUID)
  - `pipeline`: Name of the pipeline to which the version is associated
  - `version`: Name of the version of the pipeline
  - `description`: Description of the pipeline version
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the pipeline version was created
  - `last_updated`: The date when the pipeline version was last updated
  - `monitoring`: Name of a notification group which contain contacts to send monitoring notifications
  - `default_notification_group`: Name of a notification group which contain contacts to send notifications when requests for the version are completed
  - `request_retention_time`: Number of seconds to store requests to the pipeline version
  - `request_retention_mode`: Mode of request retention for requests to the pipeline version. It can be one of the following:
      - *none* - the requests will not be stored
      - *metadata* - only the metadata of the requests will be stored
      - *full* - both the metadata and input/output of the requests will be stored
  - `objects`: List of pipeline version objects
  - `attachments`: List of pipeline version object attachments
}
\description{
Get the details of a single pipeline version
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_versions_get(
   pipeline.name, version
)

# Or provide directly
result <- ubiops::pipeline_versions_get(
   pipeline.name, version,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
