% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_versions_update}
\alias{pipeline_versions_update}
\title{Update pipeline version}
\usage{
pipeline_versions_update(
  pipeline.name,
  version,
  data,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{pipeline.name}{character}

\item{version}{character}

\item{data}{named list of: [ version (optional), description (optional), labels (optional), monitoring (optional), request_retention_time (optional), request_retention_mode (optional), default_notification_group (optional), objects (optional), attachments (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created pipeline
  - `id`: Unique identifier for the pipeline version (UUID)
  - `pipeline`: Name of the pipeline to which the version is associated
  - `version`: Name of the version of the pipeline
  - `default`: A boolean indicating whether the version is the default version
  - `description`: Description of the pipeline version
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the pipeline version was created
  - `last_updated`: The date when the pipeline version was last updated
  - `monitoring`: Name of a notification group which contains contacts to send notifications when requests for the version fail and recover
  - `default_notification_group`: Name of a notification group which contains contacts to send notifications when requests for the version are completed
  - `request_retention_time`: Number of seconds to store requests to the pipeline version
  - `request_retention_mode`: Mode of request retention for requests to the pipeline version. It can be one of the following: *none*, *metadata* or *full*.
  - `objects`: List of pipeline version objects
  - `attachments`: List of pipeline version object attachments
}
\description{
Update a pipeline version. When updating labels, the labels will replace the existing value for labels. Provide the parameter 'monitoring' as the name of a notification group to send monitoring notifications to. A notification will be sent in the case of a failed/recovered request. Pass `null` to switch off monitoring notifications for this version. Provide the parameter 'default_notification_group' as the name of a notification group to send notifications when requests for the version are completed. Pass `null` to switch off request notifications for this version.  **Attention:** *In case either the `objects` or `attachments` parameter is null or an empty list, all of the objects or attachments of the pipeline will be removed.*
}
\examples{
\dontrun{
data <- list(
 version = "version",  # (optional)
 description = "description",  # (optional)
 labels = list(key = "value"),  # (optional)
 monitoring = "monitoring",  # (optional)
 request_retention_time = 0,  # (optional)
 request_retention_mode = "request_retention_mode",  # (optional)
 default_notification_group = "default_notification_group",  # (optional)
 objects = list(  # (optional)
   list(
     name = "name",
     reference_type = 'deployment',  # (optional)
     reference_name = "reference_name",
     version = "version",  # (optional)
     configuration = list(key = "value")  # (optional)
   )
 ),
 attachments = list(  # (optional)
   list(
     destination_name = "destination_name",
     sources = list(  # (optional)
       list(
         source_name = "source_name",
         mapping = list(  # (optional)
           list(
             source_field_name = source_field_name,
             destination_field_name = destination_field_name
           )
         )
       )
     )
   )
 )
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipeline_versions_update(
   pipeline.name, version, data
)

# Or provide directly
result <- ubiops::pipeline_versions_update(
   pipeline.name, version, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
