% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipelines_update}
\alias{pipelines_update}
\title{Update a pipeline}
\usage{
pipelines_update(pipeline.name, data, preload_content = TRUE, ...)
}
\arguments{
\item{pipeline.name}{character}

\item{data}{named list of: [ name (optional), description (optional), input_type (optional), input_fields (optional), output_type (optional), output_fields (optional), labels (optional), default_version (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the updated pipeline
  - `id`: Unique identifier for the pipeline (UUID)
  - `name`: Name of the pipeline
  - `project`: Project name in which the pipeline is defined
  - `description`: Description for the pipeline
  - `input_type`: Type of the pipeline input
  - `input_fields`: A list of pipeline input fields with name, data_type and widget
  - `output_type`: Type of the pipeline output
  - `output_fields`: A list of pipeline output fields with name, data_type and widget
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the pipeline was created
  - `last_updated`: The date when the pipeline was last updated
  - `default_version`: Default version of the pipeline.  If it does not have a default version, it is not set.
}
\description{
Update a pipeline. All necessary fields are validated again. When updating labels, the labels will replace the existing value for labels.
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 description = "description",  # (optional)
 input_type = "input_type",  # one of: [structured, plain]  (optional)
 input_fields = list(  # (optional)
   list(
     name = "name",
     data_type = "data_type",  # one of: [int, string, double, bool, array_int, array_double, array_string, blob, file] 
     widget = widget  # (optional)
   )
 ),
 output_type = "output_type",  # one of: [structured, plain]  (optional)
 output_fields = list(  # (optional)
   list(
     name = "name",
     data_type = "data_type",  # one of: [int, string, double, bool, array_int, array_double, array_string, blob, file] 
     widget = widget  # (optional)
   )
 ),
 labels = list(key = "value"),  # (optional)
 default_version = "default_version"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::pipelines_update(
   pipeline.name, data
)

# Or provide directly
result <- ubiops::pipelines_update(
   pipeline.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
