% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project_environment_variables_list}
\alias{project_environment_variables_list}
\title{List project environment variables}
\usage{
project_environment_variables_list(preload_content = TRUE, ...)
}
\arguments{
\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of variables described by the following fields:
  - `id`: Unique identifier for the environment variable
  - `name`: Variable name
  - `value`: Variable value (will be null for secret variables)
  - `secret`: Boolean that indicates if this variable contains sensitive information
}
\description{
List the environment variables defined for the project. These are the variables that are inherited by all deployments in this project.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::project_environment_variables_list(
   
)

# Or provide directly
result <- ubiops::project_environment_variables_list(
   
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
