% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects_list}
\alias{projects_list}
\title{List projects}
\usage{
projects_list(preload_content = TRUE, ...)
}
\arguments{
\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the projects
  - `id`: Unique identifier for the project (UUID) 
  - `name`: Name of the project 
  - `creation_date`: Time the project was created 
  - `organization_name`: Name of the organization in which the project is created
}
\description{
List all projects to which the user making request has access. The projects in organizations to which the user belongs are shown.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::projects_list(
   
)

# Or provide directly
result <- ubiops::projects_list(
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
