% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects_log_list}
\alias{projects_log_list}
\title{List logs for a project}
\usage{
projects_log_list(data = NULL, preload_content = TRUE, ...)
}
\arguments{
\item{data}{(optional) named list of: [ filters (optional), date_range (optional), date (optional), id (optional), limit (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of log details
  - `id`: Unique UUID of the log line 
  - `log`: Log line text 
  - `date`: Time the log line was created
  The following fields will be returned on response if they are set for the log line:
  - `deployment_name`:  The deployment which the log is related to 
  - `deployment_version`:  The deployment version which the log is related to 
  - `deployment_version_revision_id`: The UUID of the deployment version revision 
  - `environment_name`:  The environment which the log is related to 
  - `environment_build_id`: The UUID of the environment build 
  - `pipeline_name`: The pipeline which the log is related to 
  - `pipeline_version`: The pipeline version which the log is related to 
  - `pipeline_object_name`: The pipeline object which the log is related to 
  - `deployment_request_id`:  The deployment request the log is related to 
  - `pipeline_request_id`:  The pipeline request the log is related to 
  - `system`:  Whether the log was generated by the system (true / false) 
  - `level`: The level of the log (info / error)
}
\description{
Retrieve the logs of all objects in a project, including deployments, pipelines and requests. Using filters you can filter the logs on the objects and information of your choice.
}
\examples{
\dontrun{
data <- list(
 filters = list(key = "value"),  # (optional)
 date_range = 0,  # (optional)
 date = "date",  # (optional)
 id = "id",  # (optional)
 limit = 0  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::projects_log_list(
   
   data = NULL
)

# Or provide directly
result <- ubiops::projects_log_list(
   
   data = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
