% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects_update}
\alias{projects_update}
\title{Update a project}
\usage{
projects_update(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name (optional), advanced_permissions (optional), credits (optional), suspend (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of a project
  - `id`: Unique identifier for the project (UUID)
  - `name`: Name of the project
  - `creation_date`: Time the project was created
  - `advanced_permissions`: A boolean to enable/disable advanced permissions for the project
  - `organization_name`: Name of the organization in which the project is created
  - `credits`: Maximum usage of credits, calculated by multiplying the credit rate of a deployment instance type by the number of hours they are running
  - `suspended`: A boolean indicating whether the project is suspended due to going over the credits limit
  - `suspended_reason`: Description explaining why the project is suspended
}
\description{
Update the name of a single project. The user making the request must have appropriate permissions.
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 advanced_permissions = FALSE,  # (optional)
 credits = 0,  # (optional)
 suspend = FALSE  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::projects_update(
   data
)

# Or provide directly
result <- ubiops::projects_update(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
