% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects_usage_get}
\alias{projects_usage_get}
\title{Get resource usage}
\usage{
projects_usage_get(
  start.date = NULL,
  end.date = NULL,
  interval = "month",
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{start.date}{(optional) character}

\item{end.date}{(optional) character}

\item{interval}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 - `interval`: Interval for the usage data
  - `data_project`: A list of dictionaries containing the project usage for the given date range
  - `data_deployments`: A list of dictionaries containing the usage of each deployment in the project for the given date range
  - `data_deployments_deleted`: A list of dictionaries containing the usage corresponds to deleted deployments in the project for the given date range
}
\description{
Get credits usage for the project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::projects_usage_get(
   
   start.date = NULL, end.date = NULL, interval = 'month'
)

# Or provide directly
result <- ubiops::projects_usage_get(
   
   start.date = NULL, end.date = NULL, interval = 'month', 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
