% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_schedules.R
\name{request_schedules_create}
\alias{request_schedules_create}
\title{Create request schedules}
\usage{
request_schedules_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, object_type, object_name, version (optional), schedule, request_data (optional), timeout (optional), enabled (optional), description (optional), labels (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created request schedule
  - `name`: Name of the request 
  - `object_type`: Type of object for which the request is made 
  - `object_name`: Name of deployment/pipeline for which the request schedule is made 
  - `schedule`: Schedule in crontab format 
  - `version`: Name of version for which the request schedule is made 
  - `request_data`: Input data for the request schedule 
  - `timeout`: Timeout of the request in seconds 
  - `enabled`: Boolean value indicating whether the request schedule is enabled or disabled 
  - `creation_date`: The date when the request schedule was created 
  - `description`: Description of the request schedule 
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
}
\description{
Create a new request schedule with the provided name
}
\examples{
\dontrun{
data <- list(
 name = "name",
 object_type = "object_type",
 object_name = "object_name",
 version = "version",  # (optional)
 schedule = "schedule",
 request_data = list(key = "value"),  # (optional)
 timeout = 0,  # (optional)
 enabled = FALSE,  # (optional)
 description = "description",  # (optional)
 labels = list(key = "value")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::request_schedules_create(
   data
)

# Or provide directly
result <- ubiops::request_schedules_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
