% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_schedules.R
\name{request_schedules_list}
\alias{request_schedules_list}
\title{List request schedules}
\usage{
request_schedules_list(preload_content = TRUE, ...)
}
\arguments{
\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of all request schedules in a project
  - `name`: Name of the request 
  - `object_type`: Type of object for which the request is made 
  - `object_name`: Name of deployment/pipeline for which the request is made 
  - `schedule`: Schedule in crontab format 
  - `version`: Name of version for which the request schedule is made  
  - `request_data`: Input data for the request schedule 
  - `batch`: Boolean value indicating whether the requests will be performed as batch requests (true) or as direct requests (false) 
  - `timeout`: Timeout of the request in seconds 
  - `enabled`: Boolean value indicating whether the request schedule is enabled or disabled 
  - `creation_date`: The date when the request schedule was created 
  - `description`: Description of the request schedule 
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
}
\description{
List the request schedules in a project. The user has to have 'requests.list' permission on either 'deployments.versions' or 'pipelines.versions' to list the request schedules.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::request_schedules_list(
   
)

# Or provide directly
result <- ubiops::request_schedules_list(
   
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
