% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_schedules.R
\name{request_schedules_update}
\alias{request_schedules_update}
\title{Update a request schedule}
\usage{
request_schedules_update(schedule.name, data, preload_content = TRUE,
  ...)
}
\arguments{
\item{schedule.name}{character}

\item{data}{named list of: [ name (optional), schedule (optional), request_data (optional), batch (optional), timeout (optional), enabled (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the updated request schedule
  - `name`: Name of the request 
  - `object_type`: Type of object for which the request is made 
  - `object_name`: Name of deployment/pipeline for which the request is made 
   
  - `schedule`: Schedule in crontab format 
  - `version`: Name of version for which the request schedule is made 
  - `request_data`: Input data for the request schedule 
  - `batch`: Boolean value indicating whether the requests will be performed as batch requests (true) or as direct requests (false) 
  - `timeout`: Timeout in seconds 
  - `enabled`: Boolean value indicating whether the request schedule is enabled or disabled 
  - `creation_date`: The date when the request schedule was created
}
\description{
Update a request schedule in a project. Create permissions on the object are necessary for this action.
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 schedule = "schedule",  # (optional)
 request_data = list(key = "value"),  # (optional)
 batch = FALSE,  # (optional)
 timeout = 0,  # (optional)
 enabled = FALSE  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::request_schedules_update(
   schedule.name, data
)

# Or provide directly
result <- ubiops::request_schedules_update(
   schedule.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
