% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{revisions_list}
\alias{revisions_list}
\title{List revisions}
\usage{
revisions_list(deployment.name, version, preload_content = TRUE, ...)
}
\arguments{
\item{deployment.name}{character}

\item{version}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the revisions
  - `id`: Unique identifier for the revision (UUID)
  - `version`: Version to which the revision is linked
  - `creation_date`: The date when the revision was created
  - `created_by`: The email of the user that uploaded the deployment file. In case the revision is created by a service, the field will have a "UbiOps" value.
  - `status`: Status of the revision. Can be 'queued', 'building', 'success' or 'failed'.
  - `error_message`: Error message which explains why the revision has failed. It is empty if the revision is successful.
  - `has_request_method`: Whether the deployment code corresponding to the revision has a 'request' method
  - `has_requests_method`: Whether the deployment code corresponding to the revision has a 'requests' method
}
\description{
List all revisions associated with a version. A new revision is created every time a new deployment file is uploaded for a version.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::revisions_list(
   deployment.name, version
)

# Or provide directly
result <- ubiops::revisions_list(
   deployment.name, version,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
