% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{role_assignments_create}
\alias{role_assignments_create}
\title{Assign a role to a user in a project}
\usage{
role_assignments_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ user_id, role, object_name (optional), object_type (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created role assignment
  - `id`: Unique identifier for the role assignment (UUID) 
  - `user_id`: Unique identifier for the user (UUID) 
  - `role`: Name of the role assigned to the user 
  - `object_name`: Name of the object for which the role is assigned 
  - `object_type`: Type of the object for which the role is assigned. It can be project, deployment or pipeline.
}
\description{
Assign a role to a user in the scope of a project. This role can be assigned on either project level or on object level, which can be a deployment or pipeline. The user making the request must have appropriate permissions.
}
\examples{
\dontrun{
data <- list(
 user_id = "user_id",
 role = "role",
 object_name = "object_name",  # (optional)
 object_type = "object_type"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::role_assignments_create(
   data
)

# Or provide directly
result <- ubiops::role_assignments_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
