% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{role_assignments_per_object_list}
\alias{role_assignments_per_object_list}
\title{List roles on object/user}
\usage{
role_assignments_per_object_list(
  resource = NULL,
  resource.type = NULL,
  assignee = NULL,
  assignee.type = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{resource}{(optional) character}

\item{resource.type}{(optional) character}

\item{assignee}{(optional) character}

\item{assignee.type}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 - `id`: Unique identifier for the role assignment (UUID)
  - `assignee`: UUID of the user or the name of the object
  - `assignee_type`: Type of the assignee
  - `role`: Name of the role assigned to the user/object
  - `resource`: Name of the object for which the role is assigned
  - `resource_type`: Type of the object for which the role is assigned
}
\description{
List the roles assigned to a user or on an object in the scope of a project.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::role_assignments_per_object_list(
   
   resource = NULL, resource.type = NULL, assignee = NULL, assignee.type = NULL
)

# Or provide directly
result <- ubiops::role_assignments_per_object_list(
   
   resource = NULL, resource.type = NULL, assignee = NULL, assignee.type = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
