% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{role_assignments_per_user_list}
\alias{role_assignments_per_user_list}
\title{List the roles assigned to a specific user in a project}
\usage{
role_assignments_per_user_list(user.id, preload_content = TRUE, ...)
}
\arguments{
\item{user.id}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 - `id`: Unique identifier for the role assignment (UUID) 
  - `user_id`: Unique identifier for the user (UUID) 
  - `role`: Name of the role assigned to the user 
  - `object_name`: Name of the object for which the role is assigned 
  - `object_type`: Type of the object for which the role is assigned. It can be project, deployment or pipeline.
}
\description{
List the roles assigned to a user in the scope of a project.
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::role_assignments_per_user_list(
   user.id
)

# Or provide directly
result <- ubiops::role_assignments_per_user_list(
   user.id,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
