% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_users.R
\name{service_users_create}
\alias{service_users_create}
\title{Create a new service user}
\usage{
service_users_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name (optional), allowed_cors_origins (optional), expiry_date (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created service user
  - `id`: Unique identifier for the service user (UUID) 
  - `email`: Email of the service user  
  - `token`: The API token for the created service user  
  - `name`: Name of the service user 
  - `creation_date`: Date when the service user was created 
  - `allowed_cors_origins`: List of origin url's of which the service user is allowed to make a request from 
  - `expiry_date`: Date when the service user account will expire (UTC)
}
\description{
Create a new service user. A unique email is generated for the service user. Additionally, a token for this service user is generated. This token can be used to authorize requests sent to our API. It is possible to set an expiry date for this token. In addition, allowed cors origins can be configured for the service user. The service user will be allowed to make a deployment or pipeline request from these origins.   The token is **ONLY** returned on creation and will not be accessible afterwards.
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 allowed_cors_origins = list("value-1", "value-2")  # (optional),
 expiry_date = expiry_date  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::service_users_create(
   data
)

# Or provide directly
result <- ubiops::service_users_create(
   data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
