% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{time_series_search}
\alias{time_series_search}
\title{Search time series}
\usage{
time_series_search(
  metric = NULL,
  labels = NULL,
  custom = NULL,
  exact.match = NULL,
  limit = NULL,
  offset = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{metric}{(optional) character}

\item{labels}{(optional) character - Example: "label1:value1,label2:value2"}

\item{custom}{(optional) character ("true"|"false")}

\item{exact.match}{(optional) character ("true"|"false")}

\item{limit}{(optional) integer}

\item{offset}{(optional) integer}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of time series
  - `metric`: Name of the metric
  - `labels`: Labels that the time series has
  - `resolution`: Metric resolution in seconds
}
\description{
Search through time series
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::time_series_search(
   
   metric = NULL, labels = NULL, custom = NULL, exact.match = NULL, limit = NULL, offset = NULL
)

# Or provide directly
result <- ubiops::time_series_search(
   
   metric = NULL, labels = NULL, custom = NULL, exact.match = NULL, limit = NULL, offset = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
