% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_create}
\alias{user_create}
\title{Create a new user}
\usage{
user_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ email, password, name (optional), surname (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created user
   - `email`: Email of the user
   - `name`: Name of the user
   - `surname`: Surname of the user
}
\description{
Create a new user with the given details. After creation, an email is send to the email address to activate the account. The password needs to be at least 8 characters long.
}
\examples{
\dontrun{
data <- list(
 email = "email",
 password = "password",
 name = "name",  # (optional)
 surname = "surname"  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::user_create(
   data
)

# Or provide directly
result <- ubiops::user_create(
   data,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
