% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{webhooks_list}
\alias{webhooks_list}
\title{List webhooks}
\usage{
webhooks_list(
  labels = NULL,
  object.type = NULL,
  event = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{labels}{(optional) character - Example: "label1:value1,label2:value2"}

\item{object.type}{(optional) character}

\item{event}{(optional) character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A list of details of the webhooks in the project
  - `id`: Unique identifier for the webhook (UUID)
  - `name`: Name of the webhook
  - `url`: Callback url to send event payloads to
  - `headers`: Request headers to use when calling the webhook
  - `object_type`: Type of object for which the webhook is created
  - `object_name`: Name of deployment or pipeline for which the webhook is created
  - `version`: Name of deployment/pipeline version for which the webhook is created
  - `description`: Description of the webhook
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `event`: Event that triggers the webhook
  - `timeout`: Timeout in seconds on the call to webhook
  - `enabled`: Boolean value indicating whether the webhook is enabled or disabled
  - `retry`: Boolean value indicating whether the calls to webhook should be retried if they fail
  - `include_result`: Boolean indicating whether the result of a request should be included in the webhook call
}
\description{
List the webhooks in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::webhooks_list(
   
   labels = NULL, object.type = NULL, event = NULL
)

# Or provide directly
result <- ubiops::webhooks_list(
   
   labels = NULL, object.type = NULL, event = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
