% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{webhooks_update}
\alias{webhooks_update}
\title{Update a webhook}
\usage{
webhooks_update(webhook.name, data, preload_content = TRUE, ...)
}
\arguments{
\item{webhook.name}{character}

\item{data}{named list of: [ name (optional), url (optional), headers (optional), event (optional), description (optional), labels (optional), timeout (optional), enabled (optional), retry (optional), include_result (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of a webhook
  - `id`: Unique identifier for the webhook (UUID)
  - `name`: Name of the webhook
  - `url`: Callback url to send event payloads to
  - `headers`: Request headers to use when calling the webhook
  - `object_type`: Type of object for which the webhook is created
  - `object_name`: Name of deployment or pipeline for which the webhook is created
  - `version`: Name of deployment/pipeline version for which the webhook is created
  - `description`: Description of the webhook
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `event`: Event that triggers the webhook
  - `timeout`: Timeout in seconds on the call to webhook
  - `enabled`: Boolean value indicating whether the webhook is enabled or disabled
  - `retry`: Boolean value indicating whether the calls to webhook should be retried if they fail
  - `include_result`: Boolean indicating whether the result of a request should be included in the webhook call
}
\description{
Update a webhook
}
\examples{
\dontrun{
data <- list(
 name = "name",  # (optional)
 url = "url",  # (optional)
 headers = list(  # (optional)
   list(
     key = "key",
     value = "value",
     protected = FALSE
   )
 ),
 event = "event",  # (optional)
 description = "description",  # (optional)
 labels = list(key = "value"),  # (optional)
 timeout = 0,  # (optional)
 enabled = FALSE,  # (optional)
 retry = FALSE,  # (optional)
 include_result = FALSE  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::webhooks_update(
   webhook.name, data
)

# Or provide directly
result <- ubiops::webhooks_update(
   webhook.name, data,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
