% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response.R
\docType{class}
\name{ApiFileResponse}
\alias{ApiFileResponse}
\title{ApiFileResponse}
\format{An \code{R6Class} generator object}
\usage{
ApiFileResponse
}
\description{
ApiFileResponse Class
}
\section{Fields}{

\describe{
\item{\code{response}}{The raw response from the endpoint.}
}}

\keyword{datasets}
