% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployment_versions_get}
\alias{deployment_versions_get}
\title{Get deployment version}
\usage{
deployment_versions_get(deployment.name, version, preload_content = TRUE,
  ...)
}
\arguments{
\item{deployment.name}{character}

\item{version}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of a version
  - `id`: Unique identifier for the version (UUID)
  - `deployment`: Deployment name to which the version is associated
  - `version`: Version name
  - `description`: Description of the version
  - `language`: Language in which the version is provided
  - `status`: The status of the version
  - `active_revision`: UUID of the active revision of the version. If no deployment files have been uploaded yet, it is None.
  - `latest_build`: UUID of the latest build of the version. If no build has been triggered yet, it is None.
  - `memory_allocation`: Reserved memory for the version in MB 
  - `maximum_instances`: Upper bound of number of deployment pods running in parallel
  - `minimum_instances`: Lower bound of number of deployment pods running in parallel
  - `maximum_idle_time`: Maximum time in seconds a version stays idle before it is stopped
  - `labels`: Dictionary containing key/value pairs where key indicates the label and value is the corresponding value of that label
  - `creation_date`: The date when the version was created
  - `last_updated`: The date when the version was last updated
  - `last_file_upload`: The date when a deployment file was last uploaded for the version
  - `request_retention_time`: Number of seconds to store requests to the version
  - `request_retention_mode`: Mode of request retention for requests to the version. It can be one of the following:
      - *none* - the requests will not be stored
      - *metadata* - only the metadata of the requests will be stored
      - *full* - both the metadata and input/output of the requests will be stored
}
\description{
Retrieve details of a version of a deployment in a project
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::deployment_versions_get(
   deployment.name, version
)

# Or provide directly
result <- ubiops::deployment_versions_get(
   deployment.name, version,
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
