% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{instance_types_list}
\alias{instance_types_list}
\title{List instance types}
\usage{
instance_types_list(organization.name, preload_content = TRUE, ...)
}
\arguments{
\item{organization.name}{character}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the instance type
  - `id`: Unique identifier for the instance type (UUID) 
  - `name`: Name of the deployment instance type 
  - `display_name`: Readable name of the deployment instance type 
  - `memory_allocation`: Integer indicating memory allocation for this instance type (Mi) 
  - `cpu_allocation`: Integer indicating CPU allocation for this instance type (milliCPU) 
  - `gpu_allocation`: Integer indicating number of GPU cores for this instance type 
  - `gpu_enabled`: Boolean indicating if the GPU resource is enabled for this instance type 
  - `gpu_type`: Type of the GPU enabled for this instance type
}
\description{
Get list of available deployment instance types for an organization
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::instance_types_list(
   organization.name
)

# Or provide directly
result <- ubiops::instance_types_list(
   organization.name,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
