% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects_create}
\alias{projects_create}
\title{Create projects}
\usage{
projects_create(data, preload_content = TRUE, ...)
}
\arguments{
\item{data}{named list of: [ name, organization_name, advanced_permissions (optional), credits (optional), cors_origins (optional) ]}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 Details of the created project
  - `id`: Unique identifier for the project (UUID)
  - `name`: Name of the project
  - `creation_date`: Time the project was created
  - `advanced_permissions`: A boolean to enable/disable advanced permissions for the project
  - `organization_name`: Name of the organization in which the project is created
  - `credits`: Maximum usage of credits, calculated by multiplying the credit rate of a deployment instance type by the number of hours they are running
  - `suspended`: A boolean indicating whether the project is suspended due to going over the credits limit
  - `suspended_reason`: Description explaining why the project is suspended
  - `cors_origins`: List of origins from which the requests are allowed for the project
}
\description{
Create a new project with the provided name. **Only the organization admins have permission to make this request.** When a new project is created, the current organization admins are assigned project-admin role for the created project.
}
\examples{
\dontrun{
data <- list(
 name = "name",
 organization_name = "organization_name",
 advanced_permissions = FALSE,  # (optional)
 credits = 0,  # (optional)
 cors_origins = list("value-1", "value-2")  # (optional)
)

# Use environment variables
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::projects_create(
   data
)

# Or provide directly
result <- ubiops::projects_create(
   data,
   UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
