% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{revisions_rebuild}
\alias{revisions_rebuild}
\title{Rebuild revision}
\usage{
revisions_rebuild(
  deployment.name,
  revision.id,
  version,
  data = NULL,
  preload_content = TRUE,
  ...
)
}
\arguments{
\item{deployment.name}{character}

\item{revision.id}{character}

\item{version}{character}

\item{data}{(optional) list(key = "value")}

\item{preload_content}{(optional) Whether the API response should be preloaded. When TRUE the JSON response string is parsed to an R object. When FALSE, unprocessed API response object is returned. - Default = TRUE}

\item{...}{UBIOPS_PROJECT (system environment variable) UbiOps project name
UBIOPS_API_TOKEN (system environment variable) Token to connect to UbiOps API
UBIOPS_API_URL (optional - system environment variable) UbiOps API url - Default = "https://api.ubiops.com/v2.1"
UBIOPS_TIMEOUT (optional - system environment variable) Maximum request timeout to connect to UbiOps API - Default = NA
UBIOPS_DEFAULT_HEADERS (optional - system environment variable) Default headers to pass to UbiOps API, formatted like "header1:value1,header2:value2" - Default = ""}
}
\value{
Response from the API
 A dictionary containing details of the build
  - `id`: Unique identifier for the build (UUID)
  - `revision`: UUID of the revision to which the build is linked
  - `creation_date`: The date when the build was created
  - `status`: Status of the build. Can be 'queued', 'building', 'validating', 'success' or 'failed'.
  - `error_message`: Error message which explains why the build has failed. It is empty if the build is successful.
  - `trigger`: Action that triggered the build
  - `has_request_method`: Whether the build has a 'request' method
  - `has_requests_method`: Whether the build has a 'requests' method
}
\description{
Create a new build for a revision of a deployment version
}
\examples{
\dontrun{
# Use environment variables
Sys.setenv("UBIOPS_PROJECT" = "YOUR PROJECT NAME")
Sys.setenv("UBIOPS_API_TOKEN" = "YOUR API TOKEN")
result <- ubiops::revisions_rebuild(
   deployment.name, revision.id, version,
   data = NULL
)

# Or provide directly
result <- ubiops::revisions_rebuild(
   deployment.name, revision.id, version,
   data = NULL, 
   UBIOPS_PROJECT = "YOUR PROJECT NAME", UBIOPS_API_TOKEN = "YOUR API TOKEN"
)

print(result)

# The default API url is https://api.ubiops.com/v2.1
# Want to use a different API url?
# Provide `UBIOPS_API_URL`, either directly or as environment variable.
}
}
